[ONNX] Remove argument _retain_param_name from torch.onnx.export() function. (#61702)
As of now, the "_retain_param_name" parameter has no description in PyTorch docs website. According to code, this argument determines if we keep the original parameter names of PyTorch model in the final ONNX graph. If this is False, those original parameter names will be replaced with a series of integers starting from 1.

Since setting numbers as parameter names make no sense to users, we remove this argument from the torch.onnx.export() function to increase user experience of calling this function.

This PR will still keep it in torch.onnx.export() function for backward support while all backend logic has been changed to work as _retain_param_name is set to True.

Co-authored-by: fatcat-z <zhang-ji@outlook.com>

ghstack-source-id: b36b975857de17375ca8a6f45c7a01d9f33c171f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64370