Explicitly destory ProcessGroup in allgather_coalesced_async test
The error in #65231 means some child threads were destructed before
joined. I added some trace and prints and found that, in the failed
tests, all `assertEqual` are passed, but the `ProcessGroupGloo`
destructor wasn't called in one of the process. It could be due to
the only guarantee that Python makes is that garbage collection MAY
happen before the program exits. This commit adds an explicit
`destroy_process_group()` to alleviate the problem.

ghstack-source-id: 877444b4ef44b439b66fab6311ef716825237c42
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65513